Aplica validações e cria doc do retorno delas

+++ api/src/blls/category.js
@@ -1,7 +1,16 @@
 const Category = require('../models/category');
+const CategoryValidator = require('../validators/category');
 
 async function create(category) {
+  let errors = await CategoryValidator.validateCreate(category);
+
+  if (errors.length !== 0) {
+    return { valid: false, data: errors };
+  }
+
   await category.save();
+
+  return { valid: true, data: { _id: category._id } };
 }
 
 async function getAll() {
@@ -13,11 +22,27 @@ async function getById(id) {
 }
 
 async function remove(id) {
+  let errors = await CategoryValidator.validateRemove(id);
+
+  if (errors.length !== 0) {
+    return { valid: false, data: errors };
+  }
+
   await Category.findByIdAndRemove(id);
+
+  return { valid: true, data: { _id: id } };
 }
 
-async function update(model) {
-  await Category.findByIdAndUpdate(model._id, model);
+async function update(category) {
+  let errors = await CategoryValidator.validateUpdate(category);
+
+  if (errors.length !== 0) {
+    return { valid: false, data: errors };
+  }
+
+  await Category.findByIdAndUpdate(category._id, category);
+
+  return { valid: true, data: { _id: category._id } };
 }
 
 module.exports = { create, getAll, getById, remove, update };
